+++ backend/src/main/java/com/agrupae/infrastructure/config/ApplicationServiceConfig.java
@@ -10,7 +10,9 @@ import com.agrupae.application.port.out.course.CourseRepository;
 import com.agrupae.application.port.out.user.UserRepository;
 import com.agrupae.application.port.out.assignment.AssignmentArtifactRepository;
 import com.agrupae.application.service.assignment.AddReferenceArtifactService;
+import com.agrupae.application.service.assignment.ArchiveAssignmentService;
 import com.agrupae.application.service.assignment.CreateAssignmentService;
+import com.agrupae.application.service.assignment.EditAssignmentService;
 import com.agrupae.application.service.authentication.*;
 import com.agrupae.application.service.course.ArchiveCourseService;
 import com.agrupae.application.service.course.CreateCourseService;
@@ -92,8 +94,15 @@ public class ApplicationServiceConfig {
     }
 
     @Bean
-    public ArchiveCourseService archiveCourseService(CourseRepository courseRepository) {
-        return new ArchiveCourseService(courseRepository);
+    public ArchiveCourseService archiveCourseService(CourseRepository courseRepository,
+            AssignmentRepository assignmentRepository) {
+        return new ArchiveCourseService(courseRepository, assignmentRepository);
+    }
+
+    @Bean
+    public ArchiveAssignmentService archiveAssignmentService(AssignmentRepository assignmentRepository,
+            CourseRepository courseRepository) {
+        return new ArchiveAssignmentService(assignmentRepository, courseRepository);
     }
 
     @Bean
@@ -142,5 +151,10 @@ public class ApplicationServiceConfig {
             CourseMembershipRepository courseMembershipRepository) {
         return new GetAssignmentArtifactsService(assignmentRepository,
                 courseMembershipRepository, assignmentArtifactRepository);
+    }
+
+    public EditAssignmentService editAssignmentService(AssignmentRepository assignmentRepository,
+            CourseRepository courseRepository) {
+        return new EditAssignmentService(assignmentRepository, courseRepository);
     }
 }